Check for existing user in delete api

+++ main.py
@@ -98,6 +98,9 @@ async def update_single_user(user_id: int, user: schemas.UserUpdate, db: Session
 @app.delete("/users/{user_id}", response_model=None)
 async def delete_single_user(user_id: int, db: Session = Depends(get_db),
                              admin_user: models.Admin = Depends(get_admin)):
+    db_user = crud.get_user(db, user_id, admin_user_id=admin_user.id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     crud.delete_user(db, user_id, admin_user_id=admin_user.id)
